refactor: metric types
