+++ docs/docs/metrics/write-your-own.md
@@ -28,7 +28,8 @@ Your implementation should call `writer.write(<object_score>, <object>)` for eve
 from loguru import logger
 
 from encord_active.lib.common.iterator import Iterator
-from encord_active.lib.metrics.metric import AnnotationType, DataType, Metric, MetricType
+from encord_active.lib.metrics.metric import Metric
+from encord_active.lib.metrics.types import AnnotationType, DataType, MetricType
 from encord_active.lib.metrics.writer import CSVMetricWriter
 
 logger = logger.opt(colors=True)
