+++ docs/docs/sdk/initialize-a-project.mdx
@@ -23,7 +23,7 @@ from pathlib import Path
 from typing import List
 
 from encord_active.lib.metrics.execute import run_metrics
-from encord_active.lib.metrics.metric import EmbeddingType
+from encord_active.lib.metrics.types import EmbeddingType
 from encord_active.lib.project.local import ProjectExistsError, init_local_project
 
 # 1. Choose images to import
@@ -88,7 +88,7 @@ from pathlib import Path
 from typing import List
 
 from encord_active.lib.metrics.execute import run_metrics
-from encord_active.lib.metrics.metric import EmbeddingType
+from encord_active.lib.metrics.types import EmbeddingType
 from encord_active.lib.project.local import ProjectExistsError, init_local_project
 
 # 1. Choose images and label files to import
