+++ src/encord_active/app/model_quality/prediction_types/classification_type_builder.py
@@ -28,7 +28,7 @@ from encord_active.lib.charts.performance_by_metric import performance_rate_by_m
 from encord_active.lib.charts.scopes import PredictionMatchScope
 from encord_active.lib.embeddings.dimensionality_reduction import get_2d_embedding_data
 from encord_active.lib.embeddings.utils import Embedding2DSchema
-from encord_active.lib.metrics.metric import EmbeddingType
+from encord_active.lib.metrics.types import EmbeddingType
 from encord_active.lib.metrics.utils import MetricSchema
 from encord_active.lib.model_predictions.classification_metrics import (
     match_predictions_and_labels,
