+++ src/encord_active/lib/metrics/geometric/hu_static.py
@@ -8,12 +8,8 @@ from encord_active.lib.common.iterator import Iterator
 from encord_active.lib.common.utils import get_object_coordinates, patch_sklearn_linalg
 from encord_active.lib.embeddings.hu_moments import get_hu_embeddings
 from encord_active.lib.embeddings.writer import CSVEmbeddingWriter
-from encord_active.lib.metrics.metric import (
-    AnnotationType,
-    DataType,
-    Metric,
-    MetricType,
-)
+from encord_active.lib.metrics.metric import Metric
+from encord_active.lib.metrics.types import AnnotationType, DataType, MetricType
 from encord_active.lib.metrics.writer import CSVMetricWriter
 
 logger = logger.opt(colors=True)
