+++ src/encord_active/lib/metrics/geometric/object_size.py
@@ -8,12 +8,8 @@ from encord_active.lib.common.utils import (
     get_polygon,
 )
 from encord_active.lib.labels.object import BoxShapes, ObjectShape
-from encord_active.lib.metrics.metric import (
-    AnnotationType,
-    DataType,
-    Metric,
-    MetricType,
-)
+from encord_active.lib.metrics.metric import Metric
+from encord_active.lib.metrics.types import AnnotationType, DataType, MetricType
 from encord_active.lib.metrics.writer import CSVMetricWriter
 
 logger = logger.opt(colors=True)
