+++ src/encord_active/lib/metrics/heuristic/high_iou_changing_classes.py
@@ -2,12 +2,8 @@ from loguru import logger
 
 from encord_active.lib.common.iterator import Iterator
 from encord_active.lib.common.utils import get_iou, get_polygon
-from encord_active.lib.metrics.metric import (
-    AnnotationType,
-    DataType,
-    Metric,
-    MetricType,
-)
+from encord_active.lib.metrics.metric import Metric
+from encord_active.lib.metrics.types import AnnotationType, DataType, MetricType
 from encord_active.lib.metrics.writer import CSVMetricWriter
 
 logger = logger.opt(colors=True)
