+++ src/encord_active/lib/metrics/heuristic/missing_objects_and_wrong_tracks.py
@@ -9,12 +9,8 @@ from shapely.geometry import Polygon
 
 from encord_active.lib.common.iterator import Iterator
 from encord_active.lib.common.utils import get_iou, get_polygon
-from encord_active.lib.metrics.metric import (
-    AnnotationType,
-    DataType,
-    Metric,
-    MetricType,
-)
+from encord_active.lib.metrics.metric import Metric
+from encord_active.lib.metrics.types import AnnotationType, DataType, MetricType
 from encord_active.lib.metrics.writer import CSVMetricWriter
 
 logger = logger.opt(colors=True)
