+++ src/encord_active/lib/metrics/heuristic/object_counting.py
@@ -1,10 +1,6 @@
 from encord_active.lib.common.iterator import Iterator
-from encord_active.lib.metrics.metric import (
-    AnnotationType,
-    DataType,
-    Metric,
-    MetricType,
-)
+from encord_active.lib.metrics.metric import Metric
+from encord_active.lib.metrics.types import AnnotationType, DataType, MetricType
 from encord_active.lib.metrics.writer import CSVMetricWriter
