+++ src/encord_active/lib/metrics/semantic/_heatmap_uncertainty.py
@@ -12,13 +12,9 @@ from torchvision.io import read_image
 from torchvision.models.segmentation import DeepLabV3_MobileNet_V3_Large_Weights
 
 from encord_active.lib.common.iterator import Iterator
-from encord_active.lib.metrics.metric import (
-    AnnotationType,
-    DataType,
-    Metric,
-    MetricType,
-)
+from encord_active.lib.metrics.metric import Metric
 from encord_active.lib.metrics.semantic._class_uncertainty import train_test_split
+from encord_active.lib.metrics.types import AnnotationType, DataType, MetricType
 from encord_active.lib.metrics.writer import CSVMetricWriter
 
 DEVICE = torch.device("cuda" if torch.cuda.is_available() else "cpu")
